+++ src/replace_node.ts
@@ -1,4 +1,4 @@
-export function replace(
+export function replaceNode(
   node: ChildNode,
   ...newNodes: ReadonlyArray<string | Node>
 ): void {